% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductAnnotation.R
\name{adductAnnotation}
\alias{adductAnnotation}
\title{adductAnnotation}
\usage{
adductAnnotation(isotopeObj, PeakMtx, adductDataFrame, tolerance)
}
\arguments{
\item{isotopeObj}{List. Results from the isotopes test.}

\item{PeakMtx}{List. An rMSIprocPeakMatrix. Must contain at least the following categories: \itemize{
\item PeakMtx$intensity. A matrix containg the intensities of the peaks for each pixel (rows = pixels, cols = peaks).  
\item PeakMtx$mass. A vector containg the masses of each peak. Must be in the same order with the columns of the intensity marix.   
\item PeakMtx$numPixels. Number of pixels (rows in your matrix).   
}}

\item{adductDataFrame}{Data frame with two columns. $name, with the names of adducts to be found, $mass, with each masses.}

\item{tolerance}{Integer. Mass error tolerance in ppm.}
}
\value{
A list containing the results of the test and other kind of information.
}
\description{
Given the monoisotopic ions found by the isotopes test, founds the possible adduct pairs considering the elements in the adductDataFrame.
}
