% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{adductNetworkAnalysis}
\alias{adductNetworkAnalysis}
\title{adductNetworkAnalysis}
\usage{
adductNetworkAnalysis(
  peakMatrix,
  rMSIannotationObj,
  adductGroup = "A",
  correlationThreshold = 0.75
)
}
\arguments{
\item{peakMatrix}{An rMSIprocPeakMatrix.}

\item{rMSIannotationObj}{Output of rMSIannotation}

\item{adductGroup}{Flag to indicate from which group adduct should be used. Allowed values are "A", "B" and "AB". Refer to rMSIannotation documentation for details on each group.}

\item{correlationThreshold}{Only edges with correlation higher than the threshold will be considered for the analysis.}
}
\value{
List containing all the adduct networks found, solved or not.
}
\description{
This function reduces the adducts in tables A and/or B into clusters using network analysis.
}
