% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{AlignDataChunk}
\alias{AlignDataChunk}
\title{AlignDataChunk.}
\usage{
AlignDataChunk(rawImg, refSpectrum, startingCube = 1,
  subDataMemPercent = 10, maxcubes = 5, ...)
}
\arguments{
\item{rawImg}{The image file in custom ff data format.}

\item{refSpectrum}{The reference to align (return value of LabelFreeCreateRef()).}

\item{startingCube}{The starting ID of data cube to process}

\item{subDataMemPercent}{The maximum RAM mem usage in \% units relative to machine installed memory.}

\item{maxcubes}{limit the number of data cubes to load.}

\item{...}{}
}
\value{
The next data cube ID to be processed and the used shifts in a list
}
\description{
Aligns a Chunk of a dataset to fit the RAM machine limit. The Chunk consists in an intergar number of Cubes.
}

