% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageSpectrum.R
\name{AverageSpectrum}
\alias{AverageSpectrum}
\title{AverageSpectrum.}
\usage{
AverageSpectrum(img, NumOfThreads = parallel::detectCores())
}
\arguments{
\item{img}{An rMSI object.}

\item{NumOfThreads}{Number of threads.The Default value is the number of cores of the machine.}
}
\value{
The average spectrum.
}
\description{
Calculates the dataset average spectrum.
The average spectrum is the spectrum produced by weighting all the dataset spectra.
}
