% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deisotoper.R
\name{DeisotopingTypeCheck}
\alias{DeisotopingTypeCheck}
\title{DeisotopingTypeCheck}
\usage{
DeisotopingTypeCheck(PeakMtx, IsoNumber = 2, Tolerance = 30,
  ScoreThreshold = 0.8, ToleranceUnits = "ppm", ImageVector = NULL)
}
\description{
Checks the types of the input. Interupts the program if founds and error and give information about it.
}
