% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportROISummary.R
\name{ExportROIAverages}
\alias{ExportROIAverages}
\title{ExportROIAverages.}
\usage{
ExportROIAverages(roi_xml_file, img, out_path = getwd(), normalization = NULL)
}
\arguments{
\item{roi_xml_file}{a Bruker region XML file.}

\item{img}{an rMSI object.}

\item{out_path}{disk path where the results will be stored.}

\item{normalization}{a text string with the name of desired normalization to apply at output files.}
}
\value{
a list with the data set UUID and the pixels ID's exported for each ROI.
}
\description{
Exports the average spectrum of a group of pixels in a ROI as an ASCII text file.
ROI's are defined using Bruker XML files. A single spectrum will be exported for each ROI found in XML file.
}
