% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportROISummary.R
\name{ExportROIAveragesMultiple}
\alias{ExportROIAveragesMultiple}
\title{ExportROIAveragesMultiple.}
\usage{
ExportROIAveragesMultiple(img_list, xml_list, pkMat = NULL,
  out_path = getwd(), normalization = NULL)
}
\arguments{
\item{img_list}{a lis of rMSI objects.}

\item{xml_list}{a vector of XML ROI files in the same order as img_list.}

\item{pkMat}{an rMSIproc peak matrix object.}

\item{out_path}{disk path where the results will be stored.}

\item{normalization}{a text string with the name of desired normalization to apply at output files.}
}
\description{
Exports the ROI average spectrum for a data set of multiple images. If pkMat != NULL the CSV summary peak
matrices will be exported as well.
}
