% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{FourierBestCor}
\alias{FourierBestCor}
\title{FourierBestCor.}
\usage{
FourierBestCor(x, ref)
}
\arguments{
\item{x}{vector to correlate with ref.}

\item{ref}{Conj(FFT(ref)).}
}
\value{
the lag to shift x to fit ref with best correlation.
}
\description{
FFT Correlation
}

