% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{FourierLinearShift}
\alias{FourierLinearShift}
\title{FourierLinearShift.}
\usage{
FourierLinearShift(x, shift)
}
\arguments{
\item{x}{FFT of Data vector to be shifted in X axis direction.}

\item{shift}{The shift offset.}
}
\value{
The shifted data.
}
\description{
Time shift on a data vector using FFT.
}

