% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{FourierLinerScaleShift}
\alias{FourierLinerScaleShift}
\title{FourierLinerScaleShift.}
\usage{
FourierLinerScaleShift(x, scaling, shift)
}
\arguments{
\item{x}{the data vector to be shifted in x axis direction.}

\item{scaling}{The scaling factor ( <1 compression / >1 expansion).}

\item{shift}{The shift offset.}
}
\value{
the scaled and shifted data vector.
}
\description{
Time scale and shift on a data vector using FFT (computationally efficient).
}

