% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{FullDataSetAligment}
\alias{FullDataSetAligment}
\title{FullDataSetAligment.}
\usage{
FullDataSetAligment(raw, ...)
}
\arguments{
\item{raw}{image data in ff custom data format}

\item{...}{}
}
\value{
A list of two vectors containg the shifts performed to each spectrum. The data is stored directly in ff object.
}
\description{
TOP-METHOD: Align a whole image using multi-core capabilities and mean spectra as reference
}

