% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{GetPaddedLength}
\alias{GetPaddedLength}
\title{GetPaddedLength.}
\usage{
GetPaddedLength(N)
}
\arguments{
\item{N}{Original number of samples of data.}
}
\value{
The required number of samples to fit 2^n length for efficient FFT.
}
\description{
Get the total number of points after padding.
}

