% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsicdataio.R
\name{getrMSIdataInfo}
\alias{getrMSIdataInfo}
\title{getrMSIdataInfo.}
\usage{
getrMSIdataInfo(img)
}
\arguments{
\item{img}{an rMSI data object.}
}
\value{
a list containing all storing information.
}
\description{
Obtains all storing information of a rMSI objects and returns it as a list.
}

