% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsicdataio.R
\name{getrMSIdataInfoMultipleDataSets}
\alias{getrMSIdataInfoMultipleDataSets}
\title{getrMSIdataInfoMultipleDataSets.}
\usage{
getrMSIdataInfoMultipleDataSets(imgs_list)
}
\arguments{
\item{imgs_list}{a list of rMSI objects.}
}
\value{
a list containing all storing information unified.
}
\description{
Obtains all storing information of various rMSI objects and returns them unified in a list.
The supplied rMSI objects must have the same number of mass channels and the same data type.
Otherwise an error will be raised.
The returned list contains an extra file "datasets" to indecate at which dataset belongs each ramdisk.
}
