% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceSpectrum.R
\name{InternalReferenceSpectrum}
\alias{InternalReferenceSpectrum}
\title{InternalReferenceSpectrum.}
\usage{
InternalReferenceSpectrum(img, reference = img$mean)
}
\arguments{
\item{img}{MSI image to calculate internal reference spectrum.}

\item{reference}{the spectrum to use as reference for correlations.}
}
\value{
a list with the intensity vector corresponding to the reference spectrum and the correlation coefficient.
}
\description{
Calculates the dataset reference spectrum to use in label free alignment.
The reference spectrum is the spectrum in the dataset with the bes correlation to average spectrum.
}
