% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceSpectrum.R
\name{InternalReferenceSpectrumMultipleDatasets}
\alias{InternalReferenceSpectrumMultipleDatasets}
\title{InternalReferenceSpectrumMultipleDatasets.}
\usage{
InternalReferenceSpectrumMultipleDatasets(img_list)
}
\arguments{
\item{img_list}{a list of various rMSI objects.}
}
\value{
the intensity vector corresponding to the reference spectrum.
}
\description{
Calculates the dataset reference spectrum to use in label free alignment.
The reference spectrum is the spectrum in the dataset with the bes correlation to average spectrum.
}
