% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAnnotation.R
\name{isotopeAnnotation}
\alias{isotopeAnnotation}
\title{isotopeAnnotation}
\usage{
isotopeAnnotation(
  PeakMtx,
  isoNumber = 2,
  tolerance = 30,
  charge = 1,
  scoreThreshold = 0.8,
  toleranceUnits = "ppm",
  imageVector = NULL
)
}
\arguments{
\item{PeakMtx}{List. An rMSIprocPeakMatrix. Must contain at least the following categories: \itemize{
\item PeakMtx$intensity. A matrix containg the intensities of the peaks for each pixel (rows = pixels, cols = peaks).  
\item PeakMtx$mass. A vector containg the masses of each peak. Must be in the same order with the columns of the intensity marix.   
\item PeakMtx$numPixels. Number of pixels (rows in your matrix).   
}}

\item{isoNumber}{Integer. Number of isotopes to be found.}

\item{tolerance}{Integer. Mass tolerance for the candidates in scans or ppms.}

\item{charge}{Integer. Charge of the patterns to be found.}

\item{scoreThreshold}{Numeric. Score value to consider a ion mass a good isotope candidate. Only the ions that have this number or greater will undergo the following isotope searching stages.}

\item{toleranceUnits}{String. Must be 'ppm' or 'scan'. If ToleranceUnits is 'scan' then ImageVector must be the mass channels vector of the rMSI image (rMSIObj$mass).}

\item{imageVector}{Numeric Vector. The mass channels vector of the imaging dataset containing all the scans.}
}
\value{
A list containing the results of the test and other kind of information.
}
\description{
Finds and evaluate isotope candidates for each ion mass.
}
