% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{LabelFreeAlign}
\alias{LabelFreeAlign}
\title{LabelFreeAlign.}
\usage{
LabelFreeAlign(x, ref, limit = 200)
}
\arguments{
\item{x}{a spectrum to be aligned.}

\item{ref}{a data structre created with LabelFreeCreateRef() function.}

\item{limit}{shift limiting in ppm of spectra num of points.}
}
\value{
a vector containing: the lag_low, the lag_high and the aligned spectra to ref (spc) (all in the same vector).
}
\description{
Calculates the low and high mass shifts to perform on each side of spectrum to align it to reference.
}

