% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelFreeAlignTarImage.R
\name{LabelFreeAlign_Wizard}
\alias{LabelFreeAlign_Wizard}
\title{LabelFreeAlign_Wizard.}
\usage{
LabelFreeAlign_Wizard(storeShifts = F)
}
\arguments{
\item{storeShifts}{TRUE if applied shifts must be recorded.}
}
\value{
if storeShifts == TRUE then a list of applied shifts is returned.
}
\description{
Easy to use Label-Free aligment tool.
Just run this function and a wizard will guide you trought the alignment process.
}

