% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{LabelFreeAlignDataSet}
\alias{LabelFreeAlignDataSet}
\title{LabelFreeAlignDataSet.}
\usage{
LabelFreeAlignDataSet(ref, data, iterations = 1, multithreading = F, ...)
}
\arguments{
\item{ref}{a data structre created with LabelFreeCreateRef() function.}

\item{data}{a matrix of spectra to align.}

\item{iterations}{the number of times that alignment method must be executed.}

\item{multithreading}{TRUE if multihtreading capavilities must  be used.}

\item{...}{}
}
\value{
a matrix of spectra to align.
}
\description{
Top-Level (but not exported method) Multi Spectra aligment (threaded implemented).
}

