% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{LabelFreeCreateRef}
\alias{LabelFreeCreateRef}
\title{LabelFreeCreateRef.}
\usage{
LabelFreeCreateRef(ref, smoothing = F)
}
\arguments{
\item{ref}{reference spectrum (for example the average spectrum).}

\item{smoothing}{if TRUE, oversamples the data to achive a better alignment but a 3 times slower computations.}
}
\value{
a list containing top and bottom windowed fft.
}
\description{
Create Ref data structure to avoid unnecessary computation in ref spectrum.
}

