% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{LinearScale}
\alias{LinearScale}
\title{LinearScale.}
\usage{
LinearScale(x, scaling)
}
\arguments{
\item{x}{Data vector to be scaled in X axis direction.}

\item{scaling}{The scaling factor ( <1 compression / >1 expansion).}
}
\value{
the time scaled data.
}
\description{
Time compression/expansion using R built-in approx function (computationally sub-optimal).
}

