% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsicdataio.R
\name{loadDataCube}
\alias{loadDataCube}
\title{loadDataCube.}
\usage{
loadDataCube(img, cubeSel)
}
\arguments{
\item{img}{and rMSI object image.}

\item{cubeSel}{the cube to load R index.}
}
\value{
an R matrix with spectra in rows.
}
\description{
Loads an rMSI data cube (aka. ramdisk file) to an R matrix using C++ backend.
This method is just for testing convinience and should not be used because it copies a matrix indexed by rows to a
matrix indexed by columns so it performs very slowly. Instead of using this method use rMSI::loadImgCunckFromCube.
}

