% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoadPeakMatrixC}
\alias{LoadPeakMatrixC}
\title{LoadPeakMatrix.}
\usage{
LoadPeakMatrixC(path)
}
\arguments{
\item{path}{full path to directory from where data must be loaded.}
}
\value{
an R List containing intensity, SNR and area matrices, mass axis vector and if available the normalizations data.frame.
}
\description{
Loads a binned peaks matrix from HDD.
}

