% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{MatrixAligment}
\alias{MatrixAligment}
\title{MatrixAligment.}
\usage{
MatrixAligment(data, ...)
}
\arguments{
\item{data}{a matrix of spectra to align.}

\item{...}{iterations can be passed to specify the number of iterations over aligned data}
}
\value{
the aligned dataset in a list containing data and the applied shifts
}
\description{
TOP-METHOD: Align a whole dataset using multi-core capabilities and mean spectra as reference.
}

