% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{nextCombination}
\alias{nextCombination}
\title{nextCombination
This functions computes the following permutation of edges to obtain all the possible adduct pairs combinations in the graph}
\usage{
nextCombination(edgeCounter, numberOfMultipleEdges)
}
\description{
nextCombination
This functions computes the following permutation of edges to obtain all the possible adduct pairs combinations in the graph
}
