% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakMatrix.R
\name{OffsetRasterPosMats}
\alias{OffsetRasterPosMats}
\title{OffsetRasterPosMats.}
\usage{
OffsetRasterPosMats(posMatrix, numPixels, horizontal = T)
}
\arguments{
\item{posMatrix}{the pos matrix of an rMSIproc peak matrix object.}

\item{numPixels}{a vector with the number of pixels of each image.}

\item{horizontal}{if must be laydout horizontally or vertically.}
}
\value{
a pos Matrix with proper offsets to allow multi dataset visualization
}
\description{
creates an unified pos matrix to plot properly various datasets in a single image.
}
