% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{peakAnnotation}
\usage{
peakAnnotation(
  PeakMtx,
  iso.number = 2,
  iso.tolerance = 30,
  iso.scoreThreshold,
  iso.toleranceUnits = "ppm",
  iso.imageVector = NULL,
  add.adducts,
  add.tolerance = 10
)
}
\arguments{
\item{PeakMtx}{List. An rMSIprocPeakMatrix. Must contain at least the following categories: \itemize{
\item PeakMtx$intensity. A matrix containg the intensities of the peaks for each pixel (rows = pixels, cols = ions).  
\item PeakMtx$mass. A vector containg the masses of each peak. Must be in the same order with the columns of the intensity marix.   
\item PeakMtx$numPixels. Number of pixels (rows in your matrix).   
}}

\item{iso.number}{Integer. Number of isotopes to be found.}

\item{iso.tolerance}{Integer. Mass tolerance for the isotope candidates in scans or ppms.}

\item{iso.scoreThreshold}{Numeric. Score value to consider a ion mass a good isotope candidate. Only the ions that have this number or greater will undergo the following isotope searching stages.}

\item{iso.toleranceUnits}{String. Must be 'ppm' or 'scan'. If ToleranceUnits is 'scan' then ImageVector must be the mass channels vector of the rMSI image (rMSIObj$mass).}

\item{iso.imageVector}{Numeric Vector. The mass channels vector of the imaging dataset containing all the scans.}

\item{add.tolerance}{Integer. Mass error tolerance in ppm.}

\item{add.adductDataFrame}{Data frame with two columns.\itemize{
\item $name. Column containing the names as strings of the elements or molecules that form adducts 
\item $mass. Masses of the adduct forming elements.
}}
}
\value{
A list of lists. All the infomarion related with the isotopes can be found in $isotopes sub-list, and the same for adducts in the $adducts sub-list.
\itemize{
  \item $isotopes: List of lists. Contains the followin sub-lists:
   \itemize{
   \item $Mn: List of matrices. Contains all the M+n ions evaluated and the results matrix. The name of the lists referes to the peak selected as monoisotopic.
   \item $isotopicPeaks: Vector. Indices of the mass vector beloning to isotopic peaks.
   \item $monoisotopicPeaks: Vector. Indices of the mass vector beloning to monoisotopic peaks.
  }
  \item $adducts: List of matrices. Contains all the adduct pairs found during the test. The names refer to the neutral masses of the hypothetic ions.
}
}
\description{
Searches for isotopic ions in the peak matrix and evaluates them using morphology, intensity and mass error criteria.
This algorithm only searches carbon-based isotopic ions.
Using the infromation generated by the isotopes test, the algorithm then proceeds to search adduct pairs between the monoisotopic ions.
}
