% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{PeakList2PeakMatrix}
\alias{PeakList2PeakMatrix}
\title{PeakList2PeakMatrix.}
\usage{
PeakList2PeakMatrix(PeakList, PosMatrix, BinTolerance = 5,
  BinToleranceUsingPPM = T, BinFilter = 0.1,
  imgUUID = rMSI:::uuid_timebased(),
  imgName = "rMSIproc_PeakList2PeakMatrix_peakMat")
}
\arguments{
\item{PeakList}{R peak list. Missing values will be set to zero.}

\item{PosMatrix}{the pos matrix as following the same format as in rMSI object.}

\item{BinTolerance}{tolerance used to merge peaks to the same bin. It is recomanded to use the half of peak width in ppm units.}

\item{BinToleranceUsingPPM}{if True the peak binning tolerance is specified in ppm, if false the tolerance is set using scans.}

\item{BinFilter}{the peaks bins non detected in at least the BinFitler*TotalNumberOfPixels spectra will be deleted.}

\item{imgUUID}{a unique ID for the peak matrix, if not provided it will be automatically generated.}

\item{imgName}{the name of the original MSI dataset.}
}
\value{
peak matrix.
}
\description{
Convert's an R peak list into a peak matrix.
}
\examples{
 
  #Import an imzML using centroid mode (after peak picking).
  pkLst <- rMSIproc::import_imzMLpeakList("~/path/to/processed/data/file.imzML")
  
  #Run the peak-binning to get the peak matrix.
  myPkMat <- rMSIproc::PeakList2PeakMatrix(pkLst$peakList, pkLst$pos, BinTolerance = 25, BinToleranceUsingPPM = T)
  
  #Save the peak matrix. 
  rMSIproc::StorePeakMatrix("~/path/to/file.zip", myPkMat)

}
