% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{plotAdductNetwork}
\alias{plotAdductNetwork}
\title{plotAdductNetwork}
\usage{
plotAdductNetwork(adductNetwork, network = "solved")
}
\arguments{
\item{adductNetwork}{One of the elements returned by the adduct network analysis.}

\item{network}{String indicating which network to visualize: "complete", "solved"  or "reduced".}
}
\description{
Visualization of an adduct network
}
