% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{plotAnnotatedSpectra}
\alias{plotAnnotatedSpectra}
\title{plotAnnotatedSpectra}
\usage{
plotAnnotatedSpectra(peakMatrix, rMSIannotationObj)
}
\arguments{
\item{peakMatrix}{An rMSIprocPeakMatrix.}

\item{rMSIannotationObj}{Output of rMSIannotation}
}
\description{
Interactive visualization of the mean spectra of the peak matrix with color labaels for the monoisotopic and isotopic peaks.
}
