% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakMatrix.R
\name{plotClusterImage}
\alias{plotClusterImage}
\title{plotClusterImage.}
\usage{
plotClusterImage(peakMatrix, clusters, offset_horizontal = T, rotate = 0,
  pixel_size_um = 100)
}
\arguments{
\item{peakMatrix}{the peak matrix in an rMSIproc object.}

\item{clusters}{a vector with integer number according the cluster of each pixel.}

\item{offset_horizontal}{if true images will be laydout horizontally.}

\item{rotate}{rotation to apply.}

\item{pixel_size_um}{the pixel resolution in um.}
}
\value{
a vector with the used color for each cluster sorted according clustering numering in assending order.
}
\description{
Plot a segmentation image with the user-given clusters.
}
