% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{plotIsotopicPattern}
\alias{plotIsotopicPattern}
\title{plotIsotopicPattern}
\usage{
plotIsotopicPattern(
  peakMatrix,
  rMSIannotationObj,
  isotopicPattern,
  onlyInPattern = TRUE
)
}
\arguments{
\item{peakMatrix}{An rMSIprocPeakMatrix.}

\item{rMSIannotationObj}{The complete rMSIannotation object returned by rMSIannotation.}

\item{isotopicPattern}{An element of the list returned by rMSIannotation in the slot isotopicPatterns.}

\item{onlyInPattern}{If TRUE only the peaks in the pattern will be plotted. Otherwise, all the peaks in the mass range will be plotted}
}
\description{
Plots the mean spectra of the region close to an isotopic pattern produced by rMSIannotation
}
