% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakMatrix.R
\name{plotPeakImage}
\alias{plotPeakImage}
\title{plotPeakImage.}
\usage{
plotPeakImage(peakMatrix, mz = NULL, column = NULL, offset_horizontal = T,
  matrix = "intensity", rotate = 0, pixel_size_um = 100)
}
\arguments{
\item{peakMatrix}{the peak matrix in an rMSIproc object.}

\item{mz}{the peak mass to plot, the nearest peak mass will be ploted.}

\item{column}{the column of the peak matrix to plot.}

\item{offset_horizontal}{if true images will be laydout horizontally.}

\item{matrix}{the name of the peak matrix to plot.}

\item{rotate}{the rotation in degrees.}

\item{pixel_size_um}{the pixel resolution in um.}
}
\description{
plot the ion image map of a given mass or column of a rMSIproc peak matrix object.
If the peak matrix contains data from various datasets the images will be layout horizontally or vertically.
At leas mz or column must be specified.
}
