% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsicdataio.R
\name{printrMSIdataInfo}
\alias{printrMSIdataInfo}
\title{printrMSIdataInfo.}
\usage{
printrMSIdataInfo(img)
}
\arguments{
\item{img}{and rMSI object image.}
}
\description{
Prints HDD storing information of rMSI data object using the C++ backend.
}

