% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessImage}
\alias{ProcessImage}
\title{ProcessImage.}
\usage{
ProcessImage(
  img,
  EnableSmoothing = T,
  SmoothingKernelSize = 5,
  EnableAlignment = T,
  AlignmentIterations = 3,
  AlignmentMaxShiftppm = 200,
  AlignmentBilinear = F,
  AlignmentRefLow = 0,
  AlignmentRefMid = 0.5,
  AlignmentRefHigh = 1,
  AlignmentOversampling = 2,
  EnableCalibration = T,
  CalibrationPeakWin = 20,
  EnablePeakPicking = T,
  SNR = 5,
  peakWindow = 10,
  peakUpSampling = 10,
  UseBinning = T,
  BinTolerance = 5,
  BinFilter = 0.05,
  BinToleranceUsingPPM = F,
  EnableSpectraNormalization = T,
  EnableTICNorm = T,
  EnableRMSNorm = T,
  EnableMAXNorm = T,
  EnableTICAcqNorm = T,
  NumOfThreads = min(parallel::detectCores()/2, 6),
  CalSpan = 0.75,
  ExportPeakList = F,
  refSpc = NULL
)
}
\arguments{
\item{img}{an rMSI data object to process or a list of rMSI objects if various datasets must merged for processing.}

\item{EnableSmoothing}{a boolean declaring if smoothing alignment must be performed.}

\item{SmoothingKernelSize}{size of smoothing kernel. NULL value may be specified if EnableSmoothing is FALSE.}

\item{EnableAlignment}{a boolean declaring if label-free alignment must be performed.}

\item{AlignmentIterations}{number of iterations of label-free alignment. The rMSI ramdisk will be overwritted with aligned data. NULL value may be specified if EnableAlignment is FALSE.}

\item{AlignmentMaxShiftppm}{the maximum shift that alignment can apply in ppm. NULL value may be specified if EnableAlignment is FALSE.}

\item{AlignmentBilinear}{if TRUE the biliniar algiment mode will be used insetad of linear.}

\item{AlignmentRefLow}{the relative location of the spectrum where the bottom part correlation is calculated.}

\item{AlignmentRefMid}{the relative location of the spectrum where the central part correlation is calculated (only for bilinear mode).}

\item{AlignmentRefHigh}{the relative location of the spectrum where the top part correlation is calculated.}

\item{AlignmentOversampling}{the oversampling used in spectrum scale/shift to provide better accuracy.}

\item{EnableCalibration}{a boolean declaring if mass calibration must be performed.}

\item{CalibrationPeakWin}{the windows size used for peak detection in calibration window.}

\item{EnablePeakPicking}{a boolean declaring if peak-pickin (and binning) must be performed.}

\item{SNR}{minimal singal to noise ratio of peaks to retain. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{peakWindow}{windows size used for peak detection. Generally should be similar to peak with number of data points.  NULL value may be specified if EnablePeakPicking is FALSE.}

\item{peakUpSampling}{upsampling factor used in peak interpolation fo exact mass prediction. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{UseBinning}{if true binned matrices are returned instead of peak lists.}

\item{BinTolerance}{the tolerance used to merge peaks to the same bin. It is recomanded to use the half of peak width in ppm units. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{BinFilter}{the peaks bins non detected in at least the BinFitler*TotalNumberOfPixels spectra will be deleted. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{BinToleranceUsingPPM}{if True the peak binning tolerance is specified in ppm, if false the tolerance is set using scans.}

\item{EnableSpectraNormalization}{if normalization must be applied.}

\item{EnableTICNorm}{if TIC normalization must be performed on spectra.}

\item{EnableRMSNorm}{if RMS normalization must be performed on spectra.}

\item{EnableMAXNorm}{if MAX normalization must be performed on spectra.}

\item{EnableTICAcqNorm}{if TICAcq normalization must be performed on spectra.}

\item{NumOfThreads}{the number number of threads used to process the data.}

\item{CalSpan}{the used span for the loess fittin used in mass calibration.}

\item{ExportPeakList}{a boolean detailing wheter the un-binned peak list must be exported or not.}

\item{refSpc}{alternative reference spectrum for the alignment routine.}
}
\value{
a named list containing:
            - The process image reference (procImg).
            - The results peak-picking (peakMat). This can be returned in two forms:
                From1 (if binning is used) - a list containing three matrices (intensity, SNR and area) and a vector with a common mass axis.
                Form2 (if NO binning is applied) - a list of detected peaks for each pixel.
            - The applied mass shifts in first alignment iteration (alignShifts).
}
\description{
Perform all image pre-processing using a multi-threading implementation.
If aligment is used then the hdd files are overwirted with aligned data.
A recomeneded value of aligment ietarations is 3.
}
