% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessWizard}
\alias{ProcessWizard}
\title{ProcessWizard.}
\usage{
ProcessWizard(deleteRamdisk = T, overwriteRamdisk = F,
  calibrationSpan = 0.75, store_binsize_txt_file = F)
}
\arguments{
\item{deleteRamdisk}{if the used ramdisks for MS images must be deleted for each image processing (will be deleted after saving it to .tar file).}

\item{overwriteRamdisk}{if the current ramdisk must be overwrited.}

\item{calibrationSpan}{the used span in the loess fitting for mass calibration.}

\item{store_binsize_txt_file}{if the binSize used in each binning column must be soterd in a text file.}
}
\description{
Imports and process MSI data using a friendly GUI.
Various images can be loaded and processed with a single execution.
Data can be in XMASS, tar (rMSI) or imzML format.
Processed data will be saved in a user specified directory.
The applied processing consists in:
  - Label-free aligment (various iterations can be performed, zero iterations means no alignment).
  - Peak-picking.
  - Peak-binning.
  - Mass calibration with internal reference compounds.
Processed data includes:
  - a .tar file with the processed data.
  - a rMSIproc formated matrices with binned peaks.
  - a plain text file with used processing parameters.
}
