% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{SaveProcessingParams}
\alias{SaveProcessingParams}
\title{SaveProcessingParams.}
\usage{
SaveProcessingParams(procParams, filepath)
}
\arguments{
\item{procParams}{a list of parameters.}

\item{filepath}{a full path where params will be stored}
}
\description{
Save all parameters in a list of processing params generated using ImportWizardGui() function.
Parameters will be saved in a plain text file.
}
