% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{Smoothing}
\alias{Smoothing}
\title{Smoothing.}
\usage{
Smoothing(x, method = "SavitzkyGolay", ...)
}
\arguments{
\item{x}{the intensities of a a mass spectrum to smooth.}

\item{method}{the method to use for smoothing, available option are: "SavitzkyGolay".}

\item{...}{specific parameters to each smoothing method

Smooths a vector of data using the specified smoothing method.}
}
\value{
the smoothed data.
}
\description{
Smoothing.
}
