% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{SmoothUpper}
\alias{SmoothUpper}
\title{SmoothUpper.}
\usage{
SmoothUpper(x)
}
\arguments{
\item{x}{spectrum to smooth.}
}
\value{
smoothed data with doubled length.
}
\description{
Smooth a curve and interpolate it to double length using R built-in approx function (sub-optimal).
}

