% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{TimeWindowHigh}
\alias{TimeWindowHigh}
\title{TimeWindowHigh.}
\usage{
TimeWindowHigh(x, spectraSplit = 0.6)
}
\arguments{
\item{x}{Mass Spectra to apply window.}

\item{spectraSplit}{the high part of spectra to keep (the resting points to unit will be removed).}
}
\value{
The windowed data.
}
\description{
High-Windowing Hanning based.
}

