% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{truncateAndSolveAdductNetwork}
\alias{truncateAndSolveAdductNetwork}
\title{truncateAndSolveAdductNetwork}
\usage{
truncateAndSolveAdductNetwork(adductNetwork, vertices)
}
\arguments{
\item{adductNetwork}{An adductNetwork}

\item{vertices}{One or more integers refering to the edge to remove. The integer values of a network can be seen plotting the adductNetwork under analysis.}
}
\value{
An adductNetwork without the vertices removed.
}
\description{
Manually truncates vertices in adduct networks to solve the network. This is a manual approach, you should look for vertices that are not likly to be part of the network.
}
