% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFT_Aligment.R
\name{ZeroPadding}
\alias{ZeroPadding}
\title{ZeroPadding.}
\usage{
ZeroPadding(x, rev = F)
}
\arguments{
\item{x}{vector to be padded with zeros.}

\item{rev}{if is FALSE padd zeros at right side, else to left side.}
}
\value{
the zero padded data.
}
\description{
Zero-padding to fit 2^n length to speed-up fft.
}

