% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calibration.R
\name{CalibrationWindow}
\alias{CalibrationWindow}
\title{CalibrationWindow.}
\usage{
CalibrationWindow(mass, intensity, peak_win_size = 20, win_title = "",
  method = "loess", CalibrationSpan = 0.75)
}
\arguments{
\item{mass}{The mass vector of spectrum to calibrate.}

\item{intensity}{The intensity vector of spectrum to calibrate.}

\item{method}{a string with the method used for interpolation, valid methods are spline, linear or nocal if only target masses must be returned.}

\item{CalibrationSpan}{the span of the loess method for calibration.}

\item{use_zoo}{if the zoo package interpolation must be used}
}
\value{
a the calibrated mass axis.
}
\description{
CalibrationWindow.
}
