% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceSpectrum.R
\name{InternalReferenceSpectrum}
\alias{InternalReferenceSpectrum}
\title{InternalReferenceSpectrum.}
\usage{
InternalReferenceSpectrum(img)
}
\arguments{
\item{img}{MSI image to calculate internal reference spectrum.}
}
\value{
an intensity vector corresponding to the reference spectrum.
}
\description{
Calculates the dataset reference spectrum to use in label free alignment.
The reference spectrum is the spectrum in the dataset with the bes correlation to average spectrum.
}
