% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadStorePeakMat.R
\name{LoadPeakMatrix}
\alias{LoadPeakMatrix}
\title{LoadPeakMatrix.}
\usage{
LoadPeakMatrix(data_path)
}
\arguments{
\item{data_path}{full path to zip file where data is stored.}
}
\value{
an R List containing intensity, SNR and area matrices, mass axis vector and if available the normalizations data.frame.
}
\description{
Loads a binned peaks matrix from HDD.
}

