% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessImage}
\alias{ProcessImage}
\title{ProcessImage.}
\usage{
ProcessImage(img, EnableSmoothing = T, SmoothingKernelSize = 5,
  EnableAlignment = T, AlignmentIterations = 3,
  AlignmentMaxShiftppm = 200, EnableCalibration = T,
  CalibrationPeakWin = 20, EnablePeakPicking = T, SNR = 5,
  peakWindow = 10, peakUpSampling = 10, UseBinning = T,
  BinTolerance = 100, BinFilter = 0.05, EnableSpectraNormalization = T,
  EnableTICNorm = T, EnableMAXNorm = T, EnableTICAcqNorm = T,
  NumOfThreads = parallel::detectCores(), CalSpan = 0.75)
}
\arguments{
\item{img}{an rMSI data object to process.}

\item{EnableSmoothing}{a boolean declaring if smoothing alignment must be performed.}

\item{SmoothingKernelSize}{size of smoothing kernel. NULL value may be specified if EnableSmoothing is FALSE.}

\item{EnableAlignment}{a boolean declaring if label-free alignment must be performed.}

\item{AlignmentIterations}{number of iterations of label-free alignment. The rMSI ramdisk will be overwritted with aligned data. NULL value may be specified if EnableAlignment is FALSE.}

\item{AlignmentMaxShiftppm}{the maximum shift that alignment can apply in ppm. NULL value may be specified if EnableAlignment is FALSE.}

\item{EnableCalibration}{a boolean declaring if mass calibration must be performed.}

\item{CalibrationPeakWin}{the windows size used for peak detection in calibration window.}

\item{EnablePeakPicking}{a boolean declaring if peak-pickin (and binning) must be performed.}

\item{SNR}{minimal singal to noise ratio of peaks to retain. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{peakWindow}{windows size used for peak detection. Generally should be similar to peak with number of data points.  NULL value may be specified if EnablePeakPicking is FALSE.}

\item{peakUpSampling}{upsampling factor used in peak interpolation fo exact mass prediction. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{UseBinning}{if true binned matrices are returned instead of peak lists.}

\item{BinTolerance}{the tolerance used to merge peaks to the same bin. It is recomanded to use the half of peak width in ppm units. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{BinFilter}{the peaks bins non detected in at least the BinFitler*TotalNumberOfPixels spectra will be deleted. NULL value may be specified if EnablePeakPicking is FALSE.}

\item{EnableSpectraNormalization}{if normalization must be applied.}

\item{EnableTICNorm}{if TIC normalization must be performed on spectra.}

\item{EnableMAXNorm}{if MAX normalization must be performed on spectra.}

\item{EnableTICAcqNorm}{if TICAcq normalization must be performed on spectra.}

\item{NumOfThreads}{the number number of threads used to process the data.}

\item{CalSpan}{the used span for the loess fittin used in mass calibration.}
}
\value{
a named list containing:
            - The process image reference (procImg).
            - The results peak-picking (peakMat). This can be returned in two forms:
                From1 (if binning is used) - a list containing three matrices (intensity, SNR and area) and a vector with a common mass axis.
                Form2 (if NO binning is applied) - a list of detected peaks for each pixel.
            - The applied mass shifts in first alignment iteration (alignShifts).
}
\description{
Perform all image pre-processing using a multi-threading implementation.
If aligment is used then the hdd files are overwirted with aligned data.
A recomeneded value of aligment ietarations is 3.
}
