% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessImage}
\alias{ProcessImage}
\title{ProcessImage.}
\usage{
ProcessImage(img, performAlignment = F, SNR = 5, peakWindow = 10,
  peakUpSampling = 10, SmoothingKernelSize = 5, BinTolerance = 0.05,
  BinFilter = 0.1)
}
\arguments{
\item{img}{an rMSI data object to process.}

\item{performAlignment}{if is true FFT based aligment will be used and the ramdisk overwritted with alignment data.}

\item{SNR}{minimal singal to noise ratio of peaks to retain.}

\item{peakWindow}{windows size used for peak detection. Generally should be similar to peak with number of data points.}

\item{peakUpSampling}{upsampling factor used in peak interpolation fo exact mass prediction.}

\item{SmoothingKernelSize}{size of smoothing kernel.}
}
\value{
a intensity matrix where each row corresponds to an spectrum.
}
\description{
Perform all image pre-processing using a multi-threading implementation.
If aligment is used then the hdd files are overwirted with aligned data.
}

