% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessImage}
\alias{ProcessImage}
\title{ProcessImage.}
\usage{
ProcessImage(img, AlignmentIterations = 0, AlignmentMaxShiftppm = 200,
  SNR = 5, peakWindow = 10, peakUpSampling = 10,
  SmoothingKernelSize = 5, UseManualCalibration = T, UseBinning = T,
  BinTolerance = 0.05, BinFilter = 0.05,
  NumOfThreads = parallel::detectCores())
}
\arguments{
\item{img}{an rMSI data object to process.}

\item{AlignmentIterations}{if grather than zero FFT based aligment will be used and the ramdisk overwritted with alignment data.}

\item{AlignmentMaxShiftppm}{the maximum shift that alignment can apply in ppm.}

\item{SNR}{minimal singal to noise ratio of peaks to retain.}

\item{peakWindow}{windows size used for peak detection. Generally should be similar to peak with number of data points.}

\item{peakUpSampling}{upsampling factor used in peak interpolation fo exact mass prediction.}

\item{SmoothingKernelSize}{size of smoothing kernel, if zero smoothing is disabled.}

\item{UseManualCalibration}{if the manual calibration windows must be used.}

\item{UseBinning}{if true binned matrices are returned instead of peak lists.}

\item{BinTolerance}{the tolerance used to merge peaks to the same bin. It is recomanded to use the peak width in Da units.}

\item{BinFilter}{the peaks bins non detected in at least the BinFitler*TotalNumberOfPixels spectra will be deleted.}

\item{NumOfThreads}{the number number of threads used to process the data.}
}
\value{
a named list containing:
            - The process image reference (procImg).
            - The results peak-picking (peakMat). This can be returned in two forms:
                From1 (if binning is used) - a list containing three matrices (intensity, SNR and area) and a vector with a common mass axis.
                Form2 (if NO binning is applied) - a list of detected peaks for each pixel.
            - The applied mass shifts in first alignment iteration (alignShifts).
}
\description{
Perform all image pre-processing using a multi-threading implementation.
If aligment is used then the hdd files are overwirted with aligned data.
A recomeneded value of aligment ietarations is 3.
}

